refactor Point component, whose JSX is now re-used by multiple parent components * useImperativeHandle to access multiple refs inside the Point component
